Resize header image for srcset usage, host on a CDN
The original image is heavy (2.5mb) but it's 100vw and a major content element where quality can't be sacrificed. srcset lets us send sizes most appropriate for the user. But choosing which sizes to generate is tricky. Rather than using an abitrary set of values this adds a gulp task to generate a range of sizes with a stepped filesize savings methodology similar to responsivebreakpoints.com.

We could use Cloudinary/Imgix for the same but this avoids the proprietary dependency.

Generated images are uploaded directly to S3/CloudFront to keep them out of version control.