Fix Safari/FF issues
insertAdjacentHTML isn’t available on SVG elements in Safari. For some reason working with the nodes directly if offsetting the fill pattern, the amount seems related to “padding” added to text. Apply these values in reverse to counter.

Chrome supports iterable nodeList as of 51, FF supposedly does but 49.0 errors on QSA selection